Drop libguestfs from Requires to Suggests
This is only used for _needs_legacy_net(): It detects whether the image
needs an e1000e driver instead of virtio, i.e. whether it's RHEL/CentOS 7.

This is wrong at least for RHEL/CentOS 7.9, which work just fine with
virtio-net-pci. Thus this is only necessary for some outdated use cases,
not very interesting to current developers.

The code in instance.py already gracefully handles the absence of
libguestfs, and skips the legacy net detection then. Thus lower the
Requires: to Suggests:.

Also fix the wrong indentation of fcct.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2093717