Solves stalled image downloads
(Ab)Uses the `last_used` filed in Image to get a "liveness" status for
the image download. When the Image is in `preparing` state, the
`last_used` is touched about every 2 seconds, as long as the
`_download_remote_image` method does not stall.

The waiting processes then (can) check if the download is live/stalled,
and take over if necessary.